+++ src/translations/fr/google-recaptcha.php
@@ -41,6 +41,15 @@ return [
     'Bottom right'                                                                                                                               => 'En bas à droite',
     'Bottom left'                                                                                                                                => 'En bas à gauche',
     'Inline'                                                                                                                                     => 'En ligne',
-    'Score threshold'                                                                                                                            => 'Seuil de score',
-    'Minimum score between 0 and 1 to obtain in order for the end user to validate the reCAPTHCHA challenge. Leave blank for no score checking.' => 'Score minimum entre 0 et 1 à obtenir pour que l’utilisateur final valide le challenge reCAPTHCHA. Laissez vide pour ne pas vérifier le score.'
+    'Default Action'                                                                                                                             => 'Action par défaut',
+    'Name of the reCAPTCHA action parameter.'                                                                                                    => 'Nom du paramètre action reCAPTCHA.',
+    'Default Score Threshold'                                                                                                                    => 'Seuil de score par défaut',
+    'Score Threshold'                                                                                                                            => 'Seuil de score',
+    'Minimum score between 0 and 1 to obtain in order for the end user to validate the reCAPTHCHA challenge. Leave blank for no score checking.' => 'Score minimum entre 0 et 1 à obtenir pour que l’utilisateur final valide le challenge reCAPTHCHA. Laissez vide pour ne pas vérifier le score.',
+    'Actions Settings'                                                                                                                           => 'Paramètres Actions',
+    'Score threshold per action.'                                                                                                                => 'Seuil de score par action.',
+    'Name of the action. Only alphanumeric characters, slashes (/), and underscores (_).'                                                        => 'Nom de l’action. Uniquement caractères alphanumériques, traits bas (—) et barres obliques (/).',
+    'Minimum score to obtain in order for the end user to validate the reCAPTHCHA challenge.'                                                    => 'Score minimum à obtenir pour que l’utilisateur final valide le challenge reCAPTHCHA.',
+    'Decimal value between 0 and 1'                                                                                                              => 'Valeur décimale entre 0 et 1',
+    'Some actions values are incorrect.'                                                                                                         => 'Certaines valeurs d‘actions sont incorrectes . '
 ];